pyproject.toml
[build-system]
requires = ["setuptools>=61", "setuptools_scm[toml]>=8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nunet"
authors = [
    {name = "Seongbin Lim", email = "sungbin246@gmail.com"},
]
description = "NU-Net"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3",
  "Topic :: Scientific/Engineering",
]
dependencies = [
  "rich",
  "tqdm",
  "numpy",
  "Pillow",
  "opencv-python-headless",
  "torch",
  "torchvision",
  "tensorboard",
  "tifffile",
  "albumentations",  # scikit-image, opencv-python-headless, scipy
  "PyYAML",
  "imgaug",
  "bioimageloader",
]
dynamic = ["version"]

[project.urls]
"Homepage" = "https://github.com/LaboratoryOpticsBiosciences/nunet/"

[tool.setuptools.packages.find]
where = ["nunet"]

[tool.setuptools_scm]
